Remove mandatory dependencies on rand and num-traits crates
The `random_monge_matrix` functionality has been moved to a test
module — it's not the main focus of this crate, but it's still
convenient for us to have this functionality for tests.
