+++ benches/comparison.rs
@@ -8,6 +8,10 @@ use rand::SeedableRng;
 use rand_chacha::ChaCha20Rng;
 use test::Bencher;
 
+#[path = "../tests/random_monge/mod.rs"]
+mod random_monge;
+use random_monge::random_monge_matrix;
+
 macro_rules! repeat {
     ([ $( ($row_bench:ident, $column_bench:ident, $size:expr) $(,)* )* ],
      $row_func:path, $column_func:path) => {
@@ -15,14 +19,14 @@ macro_rules! repeat {
             #[bench]
             fn $row_bench(b: &mut Bencher) {
                 let mut rng = ChaCha20Rng::seed_from_u64(0);
-                let matrix: Array2<i32> = smawk::monge::random_monge_matrix($size, $size, &mut rng);
+                let matrix: Array2<i32> = random_monge_matrix($size, $size, &mut rng);
                 b.iter(|| $row_func(&matrix));
             }
 
             #[bench]
             fn $column_bench(b: &mut Bencher) {
                 let mut rng = ChaCha20Rng::seed_from_u64(0);
-                let matrix: Array2<i32> = smawk::monge::random_monge_matrix($size, $size, &mut rng).reversed_axes();
+                let matrix: Array2<i32> = random_monge_matrix($size, $size, &mut rng).reversed_axes();
                 b.iter(|| $column_func(&matrix));
             }
         )*
