+++ tests/complexity.rs
@@ -3,8 +3,10 @@
 use ndarray::{Array1, Array2};
 use rand::SeedableRng;
 use rand_chacha::ChaCha20Rng;
-use smawk::monge::random_monge_matrix;
 use smawk::online_column_minima;
+
+mod random_monge;
+use random_monge::random_monge_matrix;
 
 #[derive(Debug)]
 struct LinRegression {
